Updates development setup & add today's date indication
