+++ example/package.json
@@ -6,7 +6,9 @@
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-multiple-datepicker": "^0.0.2",
-    "react-scripts": "1.0.14"
+    "react-scripts": "1.0.14",
+    "styled-components": "^2.2.1",
+    "warning": "^3.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
