+++ example/src/lib/components/Datepicker/CalendarButtons.js
@@ -1,5 +1,5 @@
-import React, { Component } from 'react';
-import styled from 'styled-components';
+import React from "react";
+import styled from "styled-components";
 
 const Root = styled.div`
   display: flex;
