+++ example/src/lib/components/Datepicker/DateDisplay.js
