+++ example/src/lib/components/Datepicker/utils.js
