+++ example/src/lib/index.js
