+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "react-multiple-datepicker",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "Multiple Datepicker React",
   "main": "./build/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "cd example && yarn install && yarn start",
     "build": "./node_modules/.bin/babel ./lib -d ./build"
   },
   "repository": {
@@ -33,8 +32,6 @@
   },
   "homepage": "https://github.com/codeinfuse/react-multiple-datepicker#readme",
   "dependencies": {
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
     "styled-components": "^2.1.2",
     "warning": "^3.0.0"
   },
